GOSPODI UBEY MENYA NAHOOOOY

+++ common/models/presentation/View.php
@@ -76,7 +76,10 @@ class View extends ActiveRecord
         $view->time_answer = strtotime("now") - strtotime($start->date_start);
         $view->save(false);
 
-        $start->delete();
+        if ($start) {
+            $start->delete();
+        }
+
         $presentation = Presentation::findOne($view->presentation_id);
         $presentation->updateCounters(['views_limit' => -1]);
         foreach($answers as $answer) {
